setrace-lkm: send trace information via netlink
Adds support for sending a `setrace_record` struct (defined in
uapi/linux/setrace.h) over a generic netlink socket to userspace.
Currently an skb and genl message is built for each subscriber that is
listening on a message, this can be changed later to build the skb once
and copy it.

Additionally, move previous local SELinux function pointer declarations
to a new header (setrace_sel.h) and create a lookup table for symbols
that aren't exported, which the module needs to import.